feat: movw errorHandler to standalon file
